requirements.txt
accelerate==0.30.1
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==4.0.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
decord==0.6.0
diffusers==0.29.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fire==0.6.0
fonttools==4.57.0
fsspec==2025.3.2
gradio==5.23.3
gradio_client==1.8.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.6.1
ipython==8.35.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.8
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.11.0.86
OpenEXR==3.2.4
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pyre-extensions==0.0.29
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.4
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
spaces==0.34.1
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.3
termcolor==3.0.1
tokenizers==0.19.1
tomlkit==0.13.2
torch==2.0.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.41.2
triton==2.0.0
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
xformers==0.0.20
zipp==3.21.0
scikit-image
